Added decimal answers and UI

+++ Quick Maths/Assets/Scripts/GameManager.cs
@@ -37,8 +37,8 @@ public class GameManager : MonoBehaviour
     public static void SubmitAnswer() => OnSubmitAnswer?.Invoke();
     public static Action OnSubmitAnswer;
 
-    public static void ValidateAnswer(int answer) => OnValidateAnswer?.Invoke(answer);
-    public static Action<int> OnValidateAnswer;
+    public static void ValidateAnswer(double answer) => OnValidateAnswer?.Invoke(answer);
+    public static Action<double> OnValidateAnswer;
